// src/types/player.ts
import PrimaryPosition from "./primary-position";
import Team from "./team";

export default interface Player {
  id: number;
  fullName: string;
  link: string;
  firstName: string;
  lastName: string;
  primaryNumber: string;
  birthDate: string;
  currentAge: number;
  birthCity: string;
  birthStateProvince: string;
  birthCountry: string;
  nationality: string;
  height: string;
  weight: number;
  active: boolean;
  alternateCaptain: boolean;
  captain: boolean;
  rookie: boolean;
  shootsCatches: string;
  rosterStatus: string;
  currentTeam: {
    id: Team["id"];
    name: Team["name"];
    link: Team["link"];
  };
  primaryPosition: PrimaryPosition;
}